Use "public domain-equivalent" instead of "public"
Public licenses are not actually the same thing as
public domain-equivalent licenses. A public license is a license that
allows for more general public usage, MIT, GPL and MPL are examples of
public licenses.

Public domain-equivalent licenses, on the other hand are licenses that
provide rights equivalent to the US' public domain. So if they allow for
fully free distribution, commercial usage, modification, and relicensing
without needing to attribution the original authors of the work. IE,
they allow for the same things that they would allow for as if they were
in the public domain.